Move Main page to views folder and rename dyrectory page
