+++ FileManager/App.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using FileManager.Views;
+using System;
 using Windows.ApplicationModel;
 using Windows.ApplicationModel.Activation;
 using Windows.UI.Xaml;
