+++ FileManager/FileManager.csproj
@@ -122,17 +122,18 @@
     <Compile Include="Controlls\FileControl.xaml.cs">
       <DependentUpon>FileControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="MainPage.xaml.cs">
+    <Compile Include="Views\MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Models\FolderItem.cs" />
     <Compile Include="Models\ItemBase.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ViewModels\DirectoryFilesViewModel.cs" />
+    <Compile Include="ViewModels\Libraries\PicturesLibruaryViewModel.cs" />
     <Compile Include="ViewModels\FileControlViewModel.cs" />
+    <Compile Include="ViewModels\Libraries\LibrariesBaseViewModel.cs" />
     <Compile Include="ViewModels\MainPageViewModel.cs" />
-    <Compile Include="Views\DirectoryFilesPage.xaml.cs">
-      <DependentUpon>DirectoryFilesPage.xaml</DependentUpon>
+    <Compile Include="Views\PicturesLibruaryPage.xaml.cs">
+      <DependentUpon>PicturesLibruaryPage.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
@@ -161,11 +162,11 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="MainPage.xaml">
+    <Page Include="Views\MainPage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Views\DirectoryFilesPage.xaml">
+    <Page Include="Views\PicturesLibruaryPage.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
