--- FileManager/MainPage.xaml.cs
@@ -1,20 +1,0 @@
-﻿using FileManager.ViewModels;
-using Windows.UI.Xaml.Controls;
-
-// The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=402352&clcid=0x409
-
-namespace FileManager
-{
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
-    public sealed partial class MainPage : Page
-    {
-        public MainPageViewModel ViewModel { get; set; } = new MainPageViewModel();
-        public MainPage()
-        {
-            this.InitializeComponent();
-            this.DataContext = ViewModel;
-        }
-    }
-}
