+++ FileManager/Views/MainPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="FileManager.MainPage"
+    x:Class="FileManager.Views.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:FileManager"
