+++ FileManager/Views/PicturesLibruaryPage.xaml
@@ -28,38 +28,10 @@
                             Click="{x:Bind Path=ViewModel.GetParent}"/>
             </CommandBar.Content>
 
-
             <AppBarButton Icon="Delete" Label="Delete"/>
             <AppBarButton Icon="NewFolder" Label="New folder"/>
         </CommandBar>
 
-        <!--<StackPanel Grid.Row="0">
-            <Border>
-                <Border.Background>
-                    <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                        <GradientStop Color="#FFD96B6B"/>
-                        <GradientStop Color="#FFD96B6B" Offset="1"/>
-                        <GradientStop Color="#FFFFACAC" Offset="0.532"/>
-                    </LinearGradientBrush>
-                </Border.Background>
-                <TextBlock Text="{x:Bind Path=ViewModel.CurrentPath, Mode=TwoWay}" Margin="15" FontSize="18"/>
-            </Border>
-            <Button Content="Back" Click="{x:Bind Path=ViewModel.GetParent}"/>
-        </StackPanel>-->
-
-
-        <!--<ListBox Grid.Row="1" ItemsSource="{x:Bind Mode=TwoWay, Path=ViewModel.StorageFolders}" Margin="0,0,0,320">
-            <ListBox.ItemTemplate>
-                <DataTemplate x:Name="TripleLineDataTemplate" x:DataType="storage:IStorageItem">
-                    <StackPanel Height="100" Padding="10">
-                        <Image Source="/Images/Folder.jpg" Width="40" Height="30" Margin="0,0,10,0"/>
-                        <TextBlock Text="{x:Bind Path=Name}"  Style="{ThemeResource BaseTextBlockStyle}" Margin="0,4,0,0"/>
-                        <TextBlock Text="{x:Bind Path=Path}" Style="{ThemeResource CaptionTextBlockStyle}" Opacity=".8" Margin="0,4,0,0"/>
-                        <TextBlock Text="{x:Bind DateCreated}" Style="{ThemeResource CaptionTextBlockStyle}" Opacity=".6" Margin="0,4,0,0"/>
-                    </StackPanel>
-                </DataTemplate>
-            </ListBox.ItemTemplate>
-        </ListBox>-->
         <GridView Grid.Row="1" ItemsSource="{x:Bind ViewModel.StorageFiles, Mode=TwoWay}" DoubleTapped="{x:Bind ViewModel.OpenFolder}" SelectedItem="{x:Bind ViewModel.SelectedItem, Mode=TwoWay}" Margin="14,1,15,10">
             <GridView.ItemTemplate>
                 <DataTemplate x:DataType="viewmodels:FileControlViewModel">
